+++ locals-naming.tf
@@ -1,0 +1,7 @@
+locals {
+  # Naming locals/constants
+  name_prefix = lower(var.name_prefix)
+  name_suffix = lower(var.name_suffix)
+
+  name = coalesce(var.custom_name, azurecaf_name.keyvault.result)
+}