+++ variables.tf
@@ -23,6 +23,11 @@ variable "location" {
   type        = string
 }
 
+variable "location_short" {
+  description = "Short string for Azure location."
+  type        = string
+}
+
 variable "tenant_id" {
   description = "The Azure Active Directory tenant ID that should be used for authenticating requests to the Key Vault. Default is the current one."
   type        = string
@@ -39,12 +44,6 @@ variable "extra_tags" {
   description = "Extra tags to add"
   type        = map(string)
   default     = {}
-}
-
-variable "custom_name" {
-  description = "Name of the Key Vault, generated if not set."
-  type        = string
-  default     = ""
 }
 
 variable "enabled_for_deployment" {
@@ -95,32 +94,8 @@ variable "purge_protection_enabled" {
   default     = true
 }
 
-variable "logs_destinations_ids" {
-  type        = list(string)
-  description = "List of destination resources Ids for logs diagnostics destination. Can be Storage Account, Log Analytics Workspace and Event Hub. No more than one of each can be set. Empty list to disable logging."
-}
-
-variable "logs_categories" {
-  type        = list(string)
-  description = "Log categories to send to destinations."
-  default     = null
-}
-
-variable "logs_metrics_categories" {
-  type        = list(string)
-  description = "Metrics categories to send to destinations."
-  default     = null
-}
-
-variable "logs_retention_days" {
-  type        = number
-  description = "Number of days to keep logs on storage account"
-  default     = 30
-}
-
 variable "rbac_authorization_enabled" {
   type        = bool
   description = "Boolean flag to specify whether Azure Key Vault uses Role Based Access Control (RBAC) for authorization of data actions instead of access policies."
   default     = false
 }
-