+++ versions.tf
@@ -1,9 +1,13 @@
 terraform {
-  required_version = "> 0.12.26"
+  required_version = ">= 0.13"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
       version = ">= 2.48"
+    }
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.1"
     }
   }
 }