+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v5.0.0 - unreleased
+
+Breaking
+  * AZ-515: Option to use Azure CAF naming provider to name resources
+  * AZ-515: Require Terraform 0.13+
+
 # v4.5.0 - 2021-12-28
 
 Added